simulated processing time = 1s for each service makes total time of sequential assembly = 4s

+++ classification/src/main/java/com/hclc/enrichers/classification/contextassembler/feedback/FeedbackSummary.java
@@ -7,8 +7,8 @@ import lombok.Value;
 @Builder
 public class FeedbackSummary {
     private FeedbackType feedbackType;
-    private int numberOfAllFeedbackRequests;
-    private int numberOfIgnoredFeedbacks;
-    private int numberOfAbortedFeedbacks;
-    private int numberOfFinishedFeedbacks;
+    private long numberOfAllFeedbackRequests;
+    private long numberOfIgnoredFeedbacks;
+    private long numberOfAbortedFeedbacks;
+    private long numberOfFinishedFeedbacks;
 }
